MacOS-specific issue
When installing JDK 15 and/or 16, `/usr/bin/java`, the Kotlin plugin
uses `/usr/bin/javac` (or first version from `/usr/libexec/java_home`,
not clear which) rather than the `javac` in the current path.  This was
not noticeable with JDK 15, but with JDK 16 warnings about accessing
closed modules were turned into errors, and the build fails locally on a
MacOS.

The build is fine with Batect as that uses a JDK image which is pinned
to version 11.